requirements.txt
fastapi
uvicorn
scikit-learn
pandas
numpy
matplotlib
PyYAML
requests
python-multipart
azure-storage-blob
aiofiles
joblib